Revert PumpOpsSynchronous changes until we are finished with tests
