+++ RileyLinkKit/PumpOpsSynchronous.swift
@@ -45,18 +45,61 @@ class PumpOpsSynchronous {
     // After
     private let minimumTimeBetweenWakeAttempts = TimeInterval(minutes: 1)
     
-    var communication: PumpOpsCommunication
     let pump: PumpState
     let session: RileyLinkCmdSession
     
     init(pumpState: PumpState, session: RileyLinkCmdSession) {
         self.pump = pumpState
         self.session = session
-        self.communication = PumpOpsCommunication(session: session)
     }
     
-    internal func makePumpMessage(to messageType: MessageType, using body: MessageBody = CarelinkShortMessageBody()) -> PumpMessage {
+    private func makePumpMessage(to messageType: MessageType, using body: MessageBody = CarelinkShortMessageBody()) -> PumpMessage {
         return PumpMessage(packetType: .carelink, address: pump.pumpID, messageType: messageType, messageBody: body)
+    }
+
+    private func sendAndListen(_ msg: PumpMessage, timeoutMS: UInt16 = standardPumpResponseWindow, repeatCount: UInt8 = 0, msBetweenPackets: UInt8 = 0, retryCount: UInt8 = 3) throws -> PumpMessage {
+        let cmd = SendAndListenCmd()
+        cmd.packet = RFPacket(data: msg.txData)
+        cmd.timeoutMS = timeoutMS
+        cmd.repeatCount = repeatCount
+        cmd.msBetweenPackets = msBetweenPackets
+        cmd.retryCount = retryCount
+        cmd.listenChannel = 0
+
+        let minTimeBetweenPackets = 12 // At least 12 ms between packets for radio to stop/start
+
+        let timeBetweenPackets = max(minTimeBetweenPackets, Int(msBetweenPackets))
+
+        // 16384 = bitrate, 8 = bits per byte, 6/4 = 4b6 encoding, 1000 = ms in 1s
+        let singlePacketSendTime = (Double(msg.txData.count * 8) * 6 / 4 / 16384.0) * 1000
+
+        let totalSendTime = Double(repeatCount) * (singlePacketSendTime + Double(timeBetweenPackets))
+
+        let totalTimeout = Int(retryCount+1) * (Int(totalSendTime) + Int(timeoutMS)) + expectedMaxBLELatencyMS
+
+        guard session.doCmd(cmd, withTimeoutMs: totalTimeout) else {
+            throw PumpCommsError.rileyLinkTimeout
+        }
+
+        guard let data = cmd.receivedPacket.data else {
+            if cmd.didReceiveResponse {
+                throw PumpCommsError.unknownResponse(rx: cmd.rawReceivedData.hexadecimalString, during: "Sent \(msg)")
+            } else {
+                throw PumpCommsError.noResponse(during: "Sent \(msg)")
+            }
+        }
+
+        guard let message = PumpMessage(rxData: data) else {
+            throw PumpCommsError.unknownResponse(rx: data.hexadecimalString, during: "Sent \(msg)")
+        }
+
+        guard message.address == msg.address else {
+            throw PumpCommsError.crosstalk(message, during: "Sent \(msg)")
+        }
+
+        NotificationCenter.default.post(name: .PumpOpsSynchronousDidReceivePacket, object: self, userInfo: [type(of: self).PacketKey: cmd.receivedPacket])
+
+        return message
     }
     
     /// Attempts to send initial short wakeup message that kicks off the wakeup process.
@@ -80,14 +123,14 @@ class PumpOpsSynchronous {
             // Older pumps have a longer sleep cycle between wakeups, so send an initial burst
             do {
                 let shortPowerMessage = makePumpMessage(to: .powerOn)
-                _ = try communication.sendAndListen(shortPowerMessage, timeoutMS: 1, repeatCount: 255, msBetweenPackets: 0, retryCount: 0)
+                _ = try sendAndListen(shortPowerMessage, timeoutMS: 1, repeatCount: 255, msBetweenPackets: 0, retryCount: 0)
             }
             catch { }
         }
 
         do {
             let shortPowerMessage = makePumpMessage(to: .powerOn)
-            let shortResponse = try communication.sendAndListen(shortPowerMessage, timeoutMS: 12000, repeatCount: 255, msBetweenPackets: 0, retryCount: 0)
+            let shortResponse = try sendAndListen(shortPowerMessage, timeoutMS: 12000, repeatCount: 255, msBetweenPackets: 0, retryCount: 0)
 
             if shortResponse.messageType == .pumpAck {
                 // Pump successfully received and responded to short wakeup message!
@@ -110,7 +153,7 @@ class PumpOpsSynchronous {
     private func pumpResponding() -> Bool {
         do {
             let msg = makePumpMessage(to: .getPumpModel)
-            let response = try communication.sendAndListen(msg, retryCount: 1)
+            let response = try sendAndListen(msg, retryCount: 1)
             
             if response.messageType == .getPumpModel && response.messageBody is GetPumpModelCarelinkMessageBody {
                 return true
@@ -142,7 +185,7 @@ class PumpOpsSynchronous {
         try sendWakeUpBurst()
         
         let longPowerMessage = makePumpMessage(to: .powerOn, using: PowerOnCarelinkMessageBody(duration: duration))
-        let longResponse = try communication.sendAndListen(longPowerMessage)
+        let longResponse = try sendAndListen(longPowerMessage)
         
         guard longResponse.messageType == .pumpAck else {
             throw PumpCommsError.unexpectedResponse(longResponse, from: longPowerMessage)
@@ -158,7 +201,7 @@ class PumpOpsSynchronous {
             try wakeup()
         
             let shortMsg = makePumpMessage(to: msg.messageType)
-            let shortResponse = try communication.sendAndListen(shortMsg)
+            let shortResponse = try sendAndListen(shortMsg)
         
             guard shortResponse.messageType == .pumpAck else {
                 throw PumpCommsError.unexpectedResponse(shortResponse, from: shortMsg)
@@ -168,7 +211,7 @@ class PumpOpsSynchronous {
         }
 
         do {
-            let response = try communication.sendAndListen(msg)
+            let response = try sendAndListen(msg)
         
             guard response.messageType == responseMessageType else {
                 throw PumpCommsError.unexpectedResponse(response, from: msg)
@@ -219,7 +262,7 @@ class PumpOpsSynchronous {
         try wakeup()
         
         let msg = makePumpMessage(to: messageType)
-        let response = try communication.sendAndListen(msg)
+        let response = try sendAndListen(msg)
         
         guard response.messageType == messageType, let body = response.messageBody as? T else {
             throw PumpCommsError.unexpectedResponse(response, from: msg)
@@ -236,10 +279,10 @@ class PumpOpsSynchronous {
         
         for attempt in 0..<3 {
             do {
-                _ = try communication.sendAndListen(makePumpMessage(to: changeMessage.messageType))
+                _ = try sendAndListen(makePumpMessage(to: changeMessage.messageType))
                 
                 do {
-                    _ = try communication.sendAndListen(changeMessage, retryCount: 0)
+                    _ = try sendAndListen(changeMessage, retryCount: 0)
                 } catch {
                     // The pump does not ACK a temp basal. We'll check manually below if it was successful.
                 }
@@ -263,14 +306,14 @@ class PumpOpsSynchronous {
         try wakeup()
 
         let shortMessage = makePumpMessage(to: .changeTime)
-        let shortResponse = try communication.sendAndListen(shortMessage)
+        let shortResponse = try sendAndListen(shortMessage)
         
         guard shortResponse.messageType == .pumpAck else {
             throw PumpCommsError.unexpectedResponse(shortResponse, from: shortMessage)
         }
 
         let message = messageGenerator()
-        let response = try communication.sendAndListen(message)
+        let response = try sendAndListen(message)
         
         guard response.messageType == .pumpAck else {
             throw PumpCommsError.unexpectedResponse(response, from: message)
@@ -302,7 +345,7 @@ class PumpOpsSynchronous {
         // Identify as a MySentry device
         let findMessageResponse = PumpMessage(packetType: .mySentry, address: pump.pumpID, messageType: .pumpAck, messageBody: findMessageResponseBody)
 
-        let linkMessage = try communication.sendAndListen(findMessageResponse, timeoutMS: commandTimeoutMS)
+        let linkMessage = try sendAndListen(findMessageResponse, timeoutMS: commandTimeoutMS)
 
         guard let
             linkMessageBody = linkMessage.messageBody as? DeviceLinkMessageBody,
@@ -438,6 +481,25 @@ class PumpOpsSynchronous {
         let pumpModel = try getPumpModel()
         
         var events = [TimestampedHistoryEvent]()
+        var timeAdjustmentInterval: TimeInterval = 0
+        
+        // Going to scan backwards in time through events, so event time should be monotonically decreasing.
+        // Exceptions are Square Wave boluses, which can be out of order in the pump history by up
+        // to 8 hours on older pumps, and Normal Boluses, which can be out of order by roughly 4 minutes.
+        let eventTimestampDeltaAllowance: TimeInterval
+        if pumpModel.appendsSquareWaveToHistoryOnStartOfDelivery {
+            eventTimestampDeltaAllowance = TimeInterval(minutes: 10)
+        } else {
+            eventTimestampDeltaAllowance = TimeInterval(hours: 9)
+        }
+
+        // Start with some time in the future, to account for the condition when the pump's clock is ahead
+        // of ours by a small amount.
+        var timeCursor = Date(timeIntervalSinceNow: TimeInterval(minutes: 60))
+        
+        // Prevent returning duplicate content, which is possible e.g. in the case of rapid RF temp basal setting
+        var seenEventData = Set<Data>()
+        var lastEvent: PumpEvent?
         
         pages: for pageNum in 0..<16 {
             NSLog("Fetching page %d", pageNum)
@@ -463,76 +525,38 @@ class PumpOpsSynchronous {
             }
 
             let page = try HistoryPage(pageData: pageData, pumpModel: pumpModel)
-            
-            let (timeStampedEvents, hasMoreEvents) = convertPumpEventToTimestampedEvents(pumpEvents: page.events.reversed(), startDate: startDate, pumpModel: pumpModel)
-            
-            events = timeStampedEvents + events
-            
-            if !hasMoreEvents {
-                break
+
+            for event in page.events.reversed() {
+                if let event = event as? TimestampedPumpEvent, !seenEventData.contains(event.rawData) {
+                    seenEventData.insert(event.rawData)
+
+                    var timestamp = event.timestamp
+                    timestamp.timeZone = pump.timeZone
+
+                    if let date = timestamp.date?.addingTimeInterval(timeAdjustmentInterval) {
+                        if date.timeIntervalSince(startDate) < -eventTimestampDeltaAllowance {
+                            NSLog("Found event at (%@) to be more than %@s before startDate(%@)", date as NSDate, String(describing: eventTimestampDeltaAllowance), startDate as NSDate);
+                            break pages
+                        } else if date.timeIntervalSince(timeCursor) > eventTimestampDeltaAllowance {
+                            NSLog("Found event (%@) out of order in history. Ending history fetch.", date as NSDate)
+                            break pages
+                        } else {
+                            if (date.compare(startDate) != .orderedAscending) {
+                                timeCursor = date
+                            }
+                            events.insert(TimestampedHistoryEvent(pumpEvent: event, date: date), at: 0)
+                        }
+                    }
+                }
+
+                if let changeTimeEvent = event as? ChangeTimePumpEvent, let newTimeEvent = lastEvent as? NewTimePumpEvent {
+                    timeAdjustmentInterval += (newTimeEvent.timestamp.date?.timeIntervalSince(changeTimeEvent.timestamp.date!))!
+                }
+
+                lastEvent = event
             }
         }
         return (events, pumpModel)
-    }
-        
-    /// Converts PumpEvents after startDate to TimestampedHistoryEvents
-    /// hasMoreEvents indicates the caller should continue getting more PumpEvents from the device
-    /// - Parameters:
-    ///   - pumpEvents: array of pump events ordered from newest to oldest (reversed from normal page order)
-    ///   - startDate: return events from past this date (adjusted for TimestampDeltaAllowance)
-    ///   - pumpModel: pumpModel
-    /// - Returns: tuple of Timestamped History Events and a Bool indicating if more events can be converted
-    internal func convertPumpEventToTimestampedEvents(pumpEvents: [PumpEvent], startDate: Date, pumpModel: PumpModel) -> (events: [TimestampedHistoryEvent], hasMoreEvents: Bool) {
-        
-        // Going to scan backwards in time through events, so event time should be monotonically decreasing.
-        // Exceptions are Square Wave boluses, which can be out of order in the pump history by up
-        // to 8 hours on older pumps, and Normal Boluses, which can be out of order by roughly 4 minutes.
-        let eventTimestampDeltaAllowance: TimeInterval
-        if pumpModel.appendsSquareWaveToHistoryOnStartOfDelivery {
-            eventTimestampDeltaAllowance = TimeInterval(minutes: 10)
-        } else {
-            eventTimestampDeltaAllowance = TimeInterval(hours: 9)
-        }
-        
-        // Start with some time in the future, to account for the condition when the pump's clock is ahead
-        // of ours by a small amount.
-        var timeCursor = Date(timeIntervalSinceNow: TimeInterval(minutes: 60))
-        var events = [TimestampedHistoryEvent]()
-        var timeAdjustmentInterval: TimeInterval = 0
-        var seenEventData = Set<Data>()
-        var lastEvent: PumpEvent?
-
-        for event in pumpEvents {
-            if let event = event as? TimestampedPumpEvent, !seenEventData.contains(event.rawData) {
-                seenEventData.insert(event.rawData)
-
-                var timestamp = event.timestamp
-                timestamp.timeZone = pump.timeZone
-
-                if let date = timestamp.date?.addingTimeInterval(timeAdjustmentInterval) {
-                    if date.timeIntervalSince(startDate) < -eventTimestampDeltaAllowance {
-                        NSLog("Found event at (%@) to be more than %@s before startDate(%@)", date as NSDate, String(describing: eventTimestampDeltaAllowance), startDate as NSDate);
-                        return (events: events, hasMoreEvents: false)
-                    } else if date.timeIntervalSince(timeCursor) > eventTimestampDeltaAllowance {
-                        NSLog("Found event (%@) out of order in history. Ending history fetch.", date as NSDate)
-                        return (events: events, hasMoreEvents: false)
-                    } else {
-                        if (date.compare(startDate) != .orderedAscending) {
-                            timeCursor = date
-                        }
-                        events.insert(TimestampedHistoryEvent(pumpEvent: event, date: date), at: 0)
-                    }
-                }
-            }
-
-            if let changeTimeEvent = event as? ChangeTimePumpEvent, let newTimeEvent = lastEvent as? NewTimePumpEvent {
-                timeAdjustmentInterval += (newTimeEvent.timestamp.date?.timeIntervalSince(changeTimeEvent.timestamp.date!))!
-            }
-
-            lastEvent = event
-        }
-        
-        return (events: events, hasMoreEvents: true)
     }
     
     private func getHistoryPage(_ pageNum: Int) throws -> Data {
@@ -550,7 +574,7 @@ class PumpOpsSynchronous {
             expectedFrameNum += 1
             let msg = makePumpMessage(to: .pumpAck)
             if !curResp.lastFrame {
-                guard let resp = try? communication.sendAndListen(msg) else {
+                guard let resp = try? sendAndListen(msg) else {
                     throw PumpCommsError.rfCommsFailure("Did not receive frame data from pump")
                 }
                 guard resp.packetType == .carelink && resp.messageType == .getHistoryPage else {
@@ -662,7 +686,7 @@ class PumpOpsSynchronous {
             expectedFrameNum += 1
             let msg = makePumpMessage(to: .pumpAck)
             if !curResp.lastFrame {
-                guard let resp = try? communication.sendAndListen(msg) else {
+                guard let resp = try? sendAndListen(msg) else {
                     throw PumpCommsError.rfCommsFailure("Did not receive frame data from pump")
                 }
                 guard resp.packetType == .carelink && resp.messageType == .getGlucosePage else {
@@ -685,7 +709,7 @@ class PumpOpsSynchronous {
     
     internal func writeGlucoseHistoryTimestamp() throws -> Void {
         let shortWriteTimestamp = makePumpMessage(to: .writeGlucoseHistoryTimestamp)
-        let shortResponse = try communication.sendAndListen(shortWriteTimestamp, timeoutMS: 12000)
+        let shortResponse = try sendAndListen(shortWriteTimestamp, timeoutMS: 12000)
         
         if shortResponse.messageType == .pumpAck {
             return
@@ -738,3 +762,4 @@ public struct FrequencyScanResults {
 extension Notification.Name {
     public static let PumpOpsSynchronousDidReceivePacket = NSNotification.Name(rawValue: "com.rileylink.RileyLinkKit.PumpOpsSynchronousDidReceivePacket")
 }
+
